academic-writing/figure-format-migrator/requirements.txt
# requirements for figure-format-migrator
Pillow>=9.0.0
